split tests into individual widget files
